change buildup of page, use navigate instead of changepage
