+++ my-app/src/App.css
@@ -1,9 +1,7 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@100&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@100&family=Poiret+One&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Raleway:wght@100;200;300;400;500;600;700;800;900&display=swap");
 
 .App {
   text-align: center;
-  font-family: "Poiret One", cursive;
   flex-direction: column;
   align-self: center;
 }
