+++ my-app/src/components/general/MenuTab.tsx
@@ -1,6 +1,6 @@
-import { useEffect, useState } from "react";
-import AlphabetIcon from "./alphabetIcons/AlphabetIcon";
-import { css } from "@emotion/css";
+import {useEffect, useState} from "react";
+import AlphabetIcon from "../alphabetIcons/AlphabetIcon";
+import {css} from "@emotion/css";
 
 interface TabProps {
   tabName: string;
@@ -28,13 +28,7 @@ const MenuTab = (props: TabProps) => {
       `}
     >
       {array?.map((item) => (
-        <AlphabetIcon
-          letter={item}
-          direction="VERTICAL"
-          height={"4vw"}
-          width={"4vw"}
-          padding={"5px 0px 00px 0px"}
-        />
+        <AlphabetIcon letter={item} direction="VERTICAL" height={"4vw"} width={"4vw"} padding={"5px 0px 00px 0px"} />
       ))}
     </div>
   );
