+++ my-app/src/components/general/Socials.tsx
@@ -1,23 +1,18 @@
-import { css } from "@emotion/css";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { Resume } from "../types/typesResume";
-import {
-  faInstagram,
-  faLinkedin,
-  faGithub,
-  faFacebook,
-} from "@fortawesome/free-brands-svg-icons";
-import { Link } from "react-router-dom";
-import { IconProp, SizeProp } from "@fortawesome/fontawesome-svg-core";
-import { faLink } from "@fortawesome/fontawesome-free-solid";
-import { mergeBreakpointsInOrder } from "@mui/system";
+import {css} from "@emotion/css";
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
+import {Resume} from "../../types/typesResume";
+import {faInstagram, faLinkedin, faGithub, faFacebook} from "@fortawesome/free-brands-svg-icons";
+import {Link} from "react-router-dom";
+import {IconProp, SizeProp} from "@fortawesome/fontawesome-svg-core";
+import {faLink} from "@fortawesome/fontawesome-free-solid";
+import {mergeBreakpointsInOrder} from "@mui/system";
 
 interface Props {
   resume: Resume;
 }
 
 const Socials = (props: Props) => {
-  const { resume } = props;
+  const {resume} = props;
 
   const getIcon = (iconStr: string) => {
     switch (iconStr) {
@@ -56,15 +51,8 @@ const Socials = (props: Props) => {
               `}
               href={soc.url}
             >
-              <FontAwesomeIcon
-                icon={iconString}
-                onClick={() => {}}
-                size={"lg"}
-              />
+              <FontAwesomeIcon icon={iconString} onClick={() => {}} size={"lg"} />
             </a>
-
-            {/* 
-            <label> {soc.url}</label> */}
           </>
         );
       })}
