+++ my-app/src/components/pages/contact/Contact.tsx
@@ -1,0 +1,11 @@
+import ContactForm from "./ContactForm";
+
+const Contact = () => {
+  return (
+    <>
+      <ContactForm />
+    </>
+  );
+};
+
+export default Contact;
