--- my-app/src/styling/GlobalCSS.css
@@ -1,13 +1,0 @@
-@font-face {
-  font-family: "Raleway";
-  src: url("https://fonts.googleapis.com/css2?family=Raleway:wght@200&display=swap")
-      format("woff2"),
-    url("https://fonts.googleapis.com/css2?family=Raleway:wght@200&display=swap")
-      format("woff");
-  /* Add additional font weights/styles if needed */
-}
-
-body {
-  font-family: "Raleway", sans-serif;
-  font-size: 14px;
-}
